refactor(filerename): organise filerename
re #37
